fix haptic response when going from short to long swipe action
currently this points to the leading action. since this if statement is looking at the negative drag direction, the second bool value should be pointing at the trailing action.

I tested this and it restores a haptic response when the user swipes from a short to long gesture as expected.